pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-inspector"
version = "0.1.0"
description = "Heuristic inspector for common data file formats (prints detected type + structure)."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "pandas>=1.5",
  "openpyxl>=3.1",
  "PyYAML>=6.0",
]

[project.scripts]
data-inspector = "data_inspector.cli:main"

[tool.setuptools]
package-dir = {"" = "ingestor_island/src"}

[tool.setuptools.packages.find]
where = ["ingestor_island/src"]
